Split plugin into lib/bin for native testability
Extract all plugin logic (State, Mode, Worktree, Action, handlers,
rendering, ZellijPlugin impl, and 71 unit tests) from main.rs into
lib.rs.

main.rs is now an 11-line WASM entry point: just register_plugin!()
plus a #[cfg(not(wasm32))] stub for host_run_plugin_command (a WASM
host import that zellij-tile references, which doesn't exist on native
targets).

Why: register_plugin!() generates #[no_mangle] pub fn pipe() which
shadows libc's pipe() on native builds. This breaks any test tooling
that spawns subprocesses. Isolating it in the bin target keeps the
lib crate (and its tests) clean.

Also refactors ui.rs render functions from println!() to
writeln!(w, ...) with w: &mut impl Write, and adds render_to() on
State. This enables capturing rendered output in tests without
needing a terminal.
